Organises into packages
