+++ data/tasks.txt
@@ -1,0 +1,2 @@
+T & 1 & test6
+D & 1 & return book & 02/02/2020 1200
