+++ src/main/java/Duke.java
@@ -1,5 +1,9 @@
+import duke.util.Parser;
+import duke.util.Storage;
+import duke.util.TaskList;
+import duke.util.Ui;
+
 import java.io.IOException;
-import java.util.ArrayList;
 
 public class Duke {
     private Ui ui;
