+++ src/main/java/duke/util/DukeException.java
@@ -1,3 +1,5 @@
+package duke.util;
+
 public class DukeException extends RuntimeException{
     public DukeException(String str) {
         super(str);
