+++ src/main/java/duke/util/TaskList.java
@@ -1,3 +1,7 @@
+package duke.util;
+
+import duke.task.Task;
+
 import java.util.ArrayList;
 
 public class TaskList {
