+++ text-ui-test/EXPECTED.TXT
@@ -3,31 +3,31 @@ Hello! I'm Duke
 What can I do for you?
 ------------------------------------------------------------
 ------------------------------------------------------------
-Got it. I've added this task:
+Got it. I've added this duke.task:
   [T][ ] borrow book
-Now you have 1 task in the list.
+Now you have 1 duke.task in the list.
 ------------------------------------------------------------
 ------------------------------------------------------------
 Here are the tasks in your list:
 1.[T][ ] borrow book
 ------------------------------------------------------------
 ------------------------------------------------------------
-Got it. I've added this task:
+Got it. I've added this duke.task:
   [D][ ] return book (by: Sunday)
 Now you have 2 tasks in the list.
 ------------------------------------------------------------
 ------------------------------------------------------------
-Got it. I've added this task:
+Got it. I've added this duke.task:
   [E][ ] project meeting (at: Mon 2-4pm)
 Now you have 3 tasks in the list.
 ------------------------------------------------------------
 ------------------------------------------------------------
-Got it. I've added this task:
+Got it. I've added this duke.task:
   [D][ ] do homework (by: no idea :-p)
 Now you have 4 tasks in the list.
 ------------------------------------------------------------
 ------------------------------------------------------------
-Nice! I've marked this task as done:
+Nice! I've marked this duke.task as done:
   [D][X] return book (by: Sunday)
 ------------------------------------------------------------
 ------------------------------------------------------------
@@ -44,7 +44,7 @@ OOPS!!! The description of a event cannot be empty.
 OOPS!!! I'm sorry, but I don't know what that means :-(
 ------------------------------------------------------------
 ------------------------------------------------------------
-Noted. I've removed this task:
+Noted. I've removed this duke.task:
   [E][ ] project meeting (at: Mon 2-4pm)
 Now you have 3 tasks in the list.
 ------------------------------------------------------------
